Renamed References to Grid service to use .Service instead of .Grid. Renamed references to data module as well to match shorter name.
