+++ addon-modules/OpenSim.Data.MySQL.MoneyData/OpenSim.Data.MySQL.MoneyData/IMoneyManager.cs
@@ -31,7 +31,7 @@ using System.Collections.Generic;
 using System.Text;
 using OpenMetaverse;
 
-namespace OpenSim.Data.MySQL.MySQLMoneyDataWrapper
+namespace OpenSim.Data.MySQL.MoneyData
 {
     public interface IMoneyManager
     {
