+++ addon-modules/OpenSim.Server.MoneyServer/OpenSim.Server.MoneyServer/MoneyXmlRpcModule.cs
@@ -43,7 +43,7 @@ using OpenMetaverse;
 using OpenSim.Framework;
 using OpenSim.Framework.Servers;
 using OpenSim.Framework.Servers.HttpServer;
-using OpenSim.Data.MySQL.MySQLMoneyDataWrapper;
+using OpenSim.Data.MySQL.MoneyData;
 using OpenSim.Modules.Currency;
 using OpenSim.Region.Framework.Scenes;
 
@@ -51,7 +51,7 @@ using NSL.Network.XmlRpc;
 using NSL.Certificate.Tools;
 
 
-namespace OpenSim.Grid.MoneyServer
+namespace OpenSim.Server.MoneyServer
 {
 	class MoneyXmlRpcModule
 	{
