+++ addon-modules/OpenSim.Server.MoneyServer/OpenSim.Server.MoneyServer/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("OpenSim.Grid.MoneyServer")]
+[assembly: AssemblyTitle("OpenSim.Server.MoneyServer")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Microsoft")]
-[assembly: AssemblyProduct("OpenSim.Grid.MoneyServer")]
+[assembly: AssemblyProduct("OpenSim.Server.MoneyServer")]
 [assembly: AssemblyCopyright("Copyright © Microsoft 2009")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
